+++ CHANGELOG.md
@@ -4,7 +4,6 @@
 
 ### Added
 
-- Workflow to check linting in Python files using Pylint on push to the `main` branch.
 - Create `input` and `output` directories per each trial networks.
 - New field in `trial_network` collection called `input` that contains the inputs files sent to Jenkins.
 - New field in `trial_network` collection called `output` that contains the json received by Jenkins.
@@ -19,6 +18,7 @@
 
 - If an error occurs during TN creation, the created directory is now deleted to prevent inconsistencies.
 - When purge trial network, also remove pipelines in Jenkins used for deploy and destroy trial network.
+- Validate trial network descriptor when one of the components forming the trial network is tn_init, add tn_vxlan and tn_bastion.
 
 ### Removed